Updated so doesnt crash in android, see ARTGRAD-24
- This line would throw an exception:
  setJSBundle(instanceManager, latestJSBundleFile);

Realized in the latest source they replaced the setJSBundle function to load it differently so replaced it with latest and ran fine.
Also moved in another bug fix "clearLifecycleEventListener", which apparently prevents an infinite loop

Other than that did a diff on the files and nothing else seemed to change, except perhaps a suspend lifecycle event but not needed.

Since no issues with iOS didn't seem worth it to upgrade the entire project and risk breaking other deployments...